backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
setuptools<81
sqlalchemy==2.0.49
firebase-admin==7.4.0
google-cloud-firestore==2.27.0
python-dotenv==1.0.1
pydantic==2.12.5
pydantic[email]==2.12.5
pydantic-settings==2.13.1
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
defusedxml==0.7.1
python-multipart==0.0.26
openpyxl==3.1.5
google-genai==1.73.1
google-cloud-documentai==3.14.0
google-cloud-tasks==2.22.0
google-cloud-storage==3.10.1
python-json-logger==2.0.7
snowflake-sqlalchemy==1.6.1
snowflake-connector-python[pandas]==3.18.0
snowflake-snowpark-python==1.29.1
email-validator==2.2.0
slowapi==0.1.9
webauthn>=2.0,<3.0
asyncmy==0.2.10
alembic==1.14.0
aiosqlite==0.20.0
openai>=1.0.0
